+++ Runtime/Components/IPortableAvatarConfigTag.cs
@@ -1,0 +1,7 @@
+﻿namespace nadena.dev.ndmf.runtime.components
+{
+    internal interface IPortableAvatarConfigTag
+    {
+        
+    }
+}