+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - [#607] Fixed an issue where zh-Hans l10n text fallback to zh-Hant.
+- [#625] Null object keyframes break RewriteObjectCurves
 
 ### Changed
 - Improved the UI for Enable-Disable Plugins Window