+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Fixed
+- [#626] Fixed an issue where an exception would be thrown if the base state corresponding to a synced layer override is destroyed.
 - [#607] Fixed an issue where zh-Hans l10n text fallback to zh-Hant.
 - [#625] Null object keyframes break RewriteObjectCurves